Put all warnings in a module
