+++ openfisca_core/warnings/libyaml_warning.py
@@ -1,0 +1,5 @@
+class OpenFiscaLibyamlWarning(UserWarning):
+    """
+    Custom warning for libyaml not installed.
+    """
+    pass
