+++ openfisca_core/warnings/memory_warning.py
@@ -1,0 +1,5 @@
+class OpenFiscaMemoryWarning(UserWarning):
+    """
+    Custom warning for MemoryConfig.
+    """
+    pass
