+++ openfisca_core/warnings/tempfile_warning.py
@@ -1,0 +1,5 @@
+class OpenFiscaTempfileWarning(UserWarning):
+    """
+    Custom Warning when using temp file on disk.
+    """
+    pass
